+++ permission_handler/android/.settings/org.eclipse.buildship.core.prefs
@@ -1,2 +1,2 @@
-connection.project.dir=
+connection.project.dir=../example/android
 eclipse.preferences.version=1
